# supareel-ui/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Supareel API docs",
    "description": "API documentation of supareel platform",
    "contact": {
      "name": "Sourabh Mandal",
      "url": "https://bitsofmandal.com",
      "email": "19mandal97@gmail.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "LOCAL ENV: when your application in build and run locally"
    },
    {
      "url": "https://backend.supareel.com/api/v1",
      "description": "LOCAL ENV: when your application in build and run locally"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/upload": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "operationId": "uploadFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/{fileName}": {
      "get": {
        "tags": [
          "Uploads"
        ],
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Uploads"
        ],
        "operationId": "deleteFile",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "ERROR",
              "INFO",
              "WARNING"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "for JWT authentication",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}